target levels of support -> deltas
Instead of computing target levels of support that each elem is trying
to give, now we compute deltas to the current levels of support.

I was hoping that this would make positive feedback for support tend
to make support for the favored bindings rapidly crowd out support
for the disfavored bindings. But in fact, just as before, support
tends to oscillate and support for favored bindings doesn't tend
toward 1.0. A higher positive-feedback-rate makes the oscillations
stronger.

Hypothesis: the oscillations arise from the fact that the equilibrium
point for the levels of support given by each elem to other elems is
unstable. Now, how could I determine what that equilibrium point is,
whether it's unstable, and whether that's causing the oscillations?

A simpler approach for what to try next: Have a closer look at the
totals of the support-deltas, and see if the disfavored bindings are
getting proportionally too much support before normalization.