Antipathies no longer accumulate
...from timestep to timestep. This reduced some oscillatory behavior.

Now b->a and b->c win. Hypothesis: This is because b favors these
and b gets huge amounts of support. Support accumulates in b
because a->c and c->a get the least support (due to :want-adj-bdx)
and these are the only two bindings that don't involve b.
Consequently the other four bindings get a lot more support, and
they give some of that support back to b.