Simplified model1
The code is simpler, the algorithm is a little simpler, and the
constants are tuned reasonably.

a->b and b->c prevail, and there are only three oscillations at the
beginning, but c->a is not wiped out, and at 100 timesteps, (demo)
seems a little chaotic rather than reaching a point attractor.