test4
Tags give a basis to bindings, and support the bindings that they are
the basis of. This breaks the symmetry of preferences, and a->b
prevails. b->c leads for a while, but then the leveling induced by
diag-sigmoid equalizes them. a->b doesn't get leveled, because
diag-sigmoid doesn't reduce the leader. Maybe that should change.

In this version, bindings don't give support back to their basis tags.
That should change.