Refactored desiderata & ospecs
Now they're maps and multimethods, although the keyword ospecs still
work and are still looked up in m-desideratum-objects.

This might not have been such a good idea. I just spent hours
debugging, and the code is slower and more complicated than it
used to be. The purpose is to enable ::bdx-across-ctxs, which,
unlike the keyword ospecs, requires arguments. And the purpose
of ::bdx-across-ctxs is to enable testing that v21 can handle v20's
problem case: to avoid getting confused by the two 5's when binding
from the slipnet structure 5+6=11 to the workspace 4+5=9,6,15.
That's in test6, now under construction.