Fixed node-alias bug
try-to-bind-all was making 'succ4 bind to itself and leaving 'succ3
unbound. The bug was that graph-edit-make-node was making 'succ an
alias of 'succ3, and later 'succ4, when it created each new 'succ
tag. The fix was not to create an alias for nodes that don't have a
value.

Naming and aliasing of nodes still seems a little confused. Perhaps
it would be wiser to get rid of 'name entirely.

There is still a bug, but it's really in try-to-bind-all: when building
a 'succ node in 'workspace to correspond to a mate in 'slipnet-structure,
the new 'succ node does not get edges to the nodes that it's supposed to
tag.