Tag for equations: fills-port
:group now takes optional (:name ___) and (:class ___) arguments.

The value of a tag is the full value that you passed to add-tag to
create it, e.g. '(fill-port 5 result), rather than just '(5 result).

Variables prefixed hm- for "hash map" are now prefixed ht- for
"hash table". This seems to be the more term in Racket.

You can now pass a hash table directly to node-args->attrs.