Started breaking 'graph' into multiple units
See graph1.rkt and sigs.rkt. This idea is that you should create
the collection of graph-related functions, and the functions that
make numbo1, by linking together a bunch of units--which you can
therefore mix and match at will to run experiments.

I'm not sure yet if this is a good idea or a bad idea. So far, it
appears to be a lot of work.