bind->ht/support-given
But it's wrong. It doesn't handle the void case correctly: it
gives support. It needs to consider whether to add an edge.

The whole thing need to be rethought: building and giving support
need to be computed by the same function, and support needs to flow
into and out of edges as well as nodes.