Chipping away at fargish.rkt ...
Now when you define a nodeclass like (number n), the macro defines
a function called 'number that returns a hash table of all the
attrs for a 'number node except id. The idea is that make-node can
take this hash table as an argument instead of the present system
of passing the arguments through make-node.