simple test of rendering graph in browser
graph->jsexpr, graph->json

To run very simple demo of the force-directed graph, showing the
nodes and edges created in the set-up code in x32.rkt:

1. In one terminal window, run:

   $ racket web.rkt

2. In a web browser, open:

   http://localhost:8080/force-demo.html

To stop the web browser, hit Ctrl-C in the terminal window.

The above will probably become obsolete soon.

This is not a good layout, but it's a "spike" proving that it's
not too hard to show the graph in a browser with D3.js.