fb.rkt : Starting FARGish with br-parser-tools
I just spent the last few days struggling to understand the
documentation for brag and br-parser-tools and doing a lot of
trial-and-error experimentation.

Some non-obvious things learned:

1. If a lexer rule starts with :* (that is, allow zero or more
repetitions of something), the lexer generator will give an
undocumented warning that the lexer will accept a zero-length string.
There seems to be no way to turn off the warning. I finally figured
out that the warning means that the lexer will return a zero-length
string as a single token, indeed probably a bad thing. I thought it
was complaining that the lexer would accept a zero-length string as
the total input file.

2. The lexer will accept a token with a rule consisting of a string
literal like "word" even if there is another rule that will accept
"word" and "wordx" later (e.g. a rule that recognizes any identifier).
The first rule will not be triggered by "wordx". IOW, it does what you
want: it returns "wordx", not "word" followed by "x". The
documentation does explain this, actually, right in the second sentence
of
/Users/bkovitz/Library/Racket/7.0/pkgs/br-parser-tools-doc/br-parser-tools/doc/br-parser-tools/Lexers.html
but I wasn't expecting it there and didn't see it (technical detail
right after overall summary statement).

3. The syntactic parser generator ("parser") does not allow rules
containing string literals. So, if you want to recognize parentheses,
you can't put "(" and ")" into productions; you have to put in named
tokens like LEFT-PAREN and RIGHT-PAREN and make a separate lexer rule
for each one.

4. If you want the lexer to recognize a string literal with backslash
escapes inside it, you're going to need to put that in a separate
lexer. See get-string-token, copied from
br-parser-tools/examples/read.rkt.

There seems to be no documented way for the lexer and syntactic parser
to interact so that the lexer could allow a keyword as an identifier
in some contexts.

After spending a long time reading the brag documentation, I decided
against using it because brag requires that the grammar be separate
from the "actions" (i.e. the expressions the parser returns when it
recognizes a production), in addition to being in a separate file.
You end up having to parse your input twice: once in the parser and
again in your syntax/parse code that parses the output from the
parser.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Your branch is up to date with 'origin/master'.
#
# Changes to be committed:
#	new file:   fb.rkt
#	modified:   xlex.rkt
#
# Untracked files:
#	a
#	b.rkt
#	c.rkt
#	clojure/x.hs
#	d.rkt
#	graph1.rkt-sam
#	quick-match.rkt
#	spin.rkt
#	to-node
#	val-demo.rkt
#	xitem.rkt
#	xsupp.rkt
#