CouldMake
EquationWatcher tags operand nodes with a CouldMake node.

Did lots of refactoring. The biggest change is that now, each node's
dictionary is {'datum': o, '_hops': hop_dict} where o is a Node
object. This goes against the grain of Networkx but it seems to go
much better with the grain of Python.