EquationResultWatcher
EquationResultWatcher is a crude way to look for an opportunity
to apply an equation: it looks for the result, and if it finds it,
it sets up responses to either FillInOperator (if every operator is
Avail) or MakeSeek (the missing operators).

So far, this doesn't work very well. The model gets into a quagmire,
repeatedly MakeSeeking a number that should be marked Failed and
given up on.

To run:  py -i numbo.py