OperandView
OperandView gives support to CouldMakeFromOperands nodes that do
addition or multiplication, depending on the upper bound of what's
possible by addition. That is, if the sum of all the Avail Numbers
is less than the target, the OperandView favors multiplication.

The 'py -i numbo.py' demo can do 3 3 3; 27. I'm not sure how well
it performs on the previous easy problems.