testAutoLink now passes
I split .install_arg() into .on_init() and .on_build(). .on_init() is
called from Node.__init__(). This means that when you make a
Brick(n=2), you get an object with its 'n' attribute set to 2,
without waiting for it to be placed in the graph. And that means
that unit tests are much easier to write, e.g. testNodeWithArg.