numbo5 now recovers after overshooting in 1,1,1,1,1;4
It took some hacking to get it to work, though. Inhibition within a
sequence is now -2.0 instead of -1.0, a couple of nodeclasses are
marked is_duplicable when they shouldn't be, the recovery doesn't
happen via the slipnet, and a lot of the code is pretty error-prone,
such as making Actions vs. ActionNodes vs. doing Actions.

One unit test fails.