test_ac_selfdestruct_on_update
ActiveNodes now have an .update() method, but the system is not
satisfactory:

1. It would be nice to specify the .update action as an Ac. But then
how does it fill in missing arguments?

2. The way AcNode works now, it always returns an Action, even when
there is nothing to do. This is probably correct, since the act of
scouting for something is properly an action, not something the node
should do on each timestep simply because its .actions() method was
called; scouting should compete with other nodes' Actions. But for
.update(), we don't want to return an Action if there is nothing to
do.

In the current unit test, AllBricksAvail.update() does not use Ac.