.update with Acs; MyContext
Now AllBricksAvail in testAc.py specifies its update Action with
Acs. The problem of passing an argument for 'within' is solved by
setting 'within' to MyContext within the All Ac. Ac.get() knows to
translate a MyContext object to the group node that contains the
actor.

This is still not satisfactory because AllBricksAvail now returns an
update Action every timestep. This might be OK for AllBricksAvail
(though it's a little inefficient), but unconditionally returning
an update Action would make a genuine ActiveNode do nothing but
updates and never its main .actions().