test_ac_fillparamscout
Quote('string') enables AcEnv.get() to return strings, inhibiting
recursive lookup.

LookForArg is unsatisfactory: it assumes that the type of node to be
searched for is Glom. The object of the search should depend on the
name of the parameter to be filled, and perhaps on more (like elements
in the surrounding context, including what has been tried before).