problem -- problem_solver
This is a band-aid for a bug-prone aspect of the system: it will treat
nodeids, lists of nodeids, sets of nodeids, etc. as integers, etc. if
you pass them to .add_node() as mates for a port label that isn't
defined in .port_mates. Consequently, you can build multiple,
"identical" nodes, such as LookForOperand in the previous version of
numbo6.py, if the nodes' mates are specified in different ways with
equivalent meaning, e.g. problem=25, problem={25},
problem=g.as_node(25).