New way to run a PassiveChain
Now PassiveChain starts a LiveActiveChain containing all needed active
nodes at once. Each active node must do a FindBasis() successfully
before it can do its real Action. So, the active nodes might run out of
sequence, but they won't get anywhere until the nodes they depend on are
built--which might include tags that demonstrate that a condition is
satisfied, like DiffIsWanted.

A big advantage of this approach is that no PassiveChainRunner is needed
to monitor the live passive nodes as they're built. This avoids the very
nasty problem of determining what counts as a completing the next step
in the chain, as was done via the 'completion' port in the previous
check-in. In the new way, all the live active nodes just run whenever
they can. Each individual active node "knows" what it needs to get
started.