Mutual antipathy between Consumes
The model can now, pretty much, solve the pons asinorum: 4 5 6; 15.

It's still pretty ugly:
The __main__ in FARGish2.py gives the model some extra priming at the
start of the run. You have to run g.do_timestep() about 10 iterations
to see the the 9+6=15 get painted (probably, not certainly), and then
another few iterations for the Detector to notice that the numble has
been solved an raise an exception.

There are still many debugging statements.