Fixed a bug; rm early binding in replace_refs()
The bug was the early binding done by replace_refs(). It caused the
behalf_of argument in the Paint codelet to be set to None when
the BuildLitPainter codelet was run without an agent. When the
resulting LitPainter agent ran, its Paint codelet's behalf_of
overrode the BehalfOf object at the end of the sources list.
Consequently, the Paint codelet's NewState('behalf_of', Succeeded)
codelet had no reference to the LitPainter that ran it.

The fix was to make running replace_refs() on an Agent refrain from
replacing the Refs inside the Agent's codelets. (Was it ever a good
idea to put Refs inside an Agent? Do we need to run replace_refs()
on Agents?) Now, run_one_codelet() always calls replace_refs() on a
codelet before running it. Possibly that wastes CPU time, but oh well.

There was another aspect to this bug: why did Paint have a behalf_of
inside it anyway? behalf_of refers to the agent that is actually running
the codelet. So, I removed behalf_of from the Paint codelet's fields
but left it in its run() method. Possibly even this isn't right; we
might someday need to specify that a codelet is to be run on behalf
of a different agent than the one that ran it.

Finding the bug was hard. I added the @trace decorator to util.py. It
proved its worth immediately.

Fixing that bug exposed that a couple of tests did not pass the agent
when calling run_codelet(). Fixed.