Bug fix: RMemAbs takes termination_threshold
The bug was that util.fields_for() did not recognize that the RMemAbs
class (and many others) had a termination_threshold field that could be
given a value in the constructor. The fix was to declare RMemAbs a
dataclass (with the @dataclass decorator) and, when dynamically
creating subclasses of RMem, to create them by calling make_dataclass()
(now in RMem.make_class()) rather than by calling the three-argument
form of type().