New conversation with Abhinav
New ideas:

1. Instead of figuring out "Given K, what should P be?" etc. within
the loop that steps through all the predicates, that loop returns an
IncompleteProposal object possibly containing LoopVariables. These
specify a new loop, in which we evaluate variables inside the
ActionElements to make determinate ActionProposals.

2. p5 creates an IncompletePainter object, which sits in the workspace
and gets converted, possibly in multiple, competing ways, into a
determinate painter.

3. ConsecutiveLocations should refer only to snippets. This way,
the IncompletePainter created by p5, and any further painters that it
creates, can be bound to or limited to a new snippet.