test_painter_cluster_reoeat() now passes,
with a test that only one Succ object was created, i.e. no F1=Succ and
F2=Succ. I suspect, though, that the code to exit from a PainterCluster
is still not thought through. It needs to be refactored and then
examined to verify that it won't get into an infinite loop and that each
type of variable or object in the local_ws gets mapped to something
appropriate in the main workspace.